Update link and description in footer
The hyperlink and description in the footer of index.html have been updated. The link that was previously for the ESP32 Radar installer has been switched to FormationFlight and now points to the updated Github page. Concurrently, the footer text has been changed from "ESP32 Radar Web Install" to "FormationFlight Web Install".