Update update.py to install requirements
In the `update` function of update.py, we replaced an infinite loop with a command to "pip install -r requirements.txt". We've also adjusted the invocation of `download_latest_release_bin_assets` and `generate_manifests` to run only once. This revision ensures smoother software deployment by installing necessary Python packages and executing the scripts only as much as needed.