Update version number and simplify device labeling in generate_manifests.py
The script's version number has been updated from "3.02" to "5.0". Changes have been made to simplify the extraction process for creating device labels, which was previously complex and more susceptible to errors. If a label can't be generated, the script is now set to return the file name as a fallback.