Improve HTML structure of index.html
This commit modifies the HTML structure of the index.html file to improve readability and layout. Changes include adjustments in indentation, reformatting long lines of text, and removing comment on fetching and displaying release notes which lead to cleaner code. The cleaner structure leads to easier future modifications and readability.