Remove release command from Procfile and add it to postdeploy script.
The 'release' command was removed from the Procfile, and instead added to the 'postdeploy' script in the app.json file. This modification ensures that the Procfile is copied into the /app directory after the application is deployed, ensuring it is in the correct location for the continued operation of the application.