Update loading release version text style
The "Work In Progress. . . USE AT Your Own Risk . . ." message was removed from index.html and the font style for "Loading Release Version" was updated. Instead of just displaying the text, it's now presented in a slightly larger and bolder font to provide a cleaner look to the webpage and improve readability.